wahoo: anykernel.sh: Detect if user is in recovery or not
We don't need to mount /system or do any weird stuff with LD variables
when not in recovery.

+++ anykernel.sh
@@ -22,6 +22,12 @@ is_slot_device=1;
 ramdisk_compression=auto;
 
 
+# Detect whether we're in recovery or booted up
+ps | grep zygote | grep -v grep >/dev/null && in_recovery=false || in_recovery=true;
+! $in_recovery || ps -A 2>/dev/null | grep zygote | grep -v grep >/dev/null && in_recovery=false;
+! $in_recovery || id | grep -q 'uid=0' || in_recovery=false;
+
+
 ## AnyKernel methods (DO NOT CHANGE)
 # import patching functions/variables - see for reference
 . /tmp/anykernel/tools/ak2-core.sh;
@@ -37,12 +43,14 @@ chown -R root:root $ramdisk/*;
 split_boot;
 
 
-# Mount system to get some information about the user's setup
-umount /system;
-umount /system 2>/dev/null;
-mkdir /system_root 2>/dev/null;
-mount -o ro -t auto /dev/block/bootdevice/by-name/system$slot /system_root;
-mount -o bind /system_root/system /system;
+# Mount system to get some information about the user's setup (only needed in recovery)
+if $in_recovery; then
+  umount /system;
+  umount /system 2>/dev/null;
+  mkdir /system_root 2>/dev/null;
+  mount -o ro -t auto /dev/block/bootdevice/by-name/system$slot /system_root;
+  mount -o bind /system_root/system /system;
+fi;
 
 
 # Warn user of their support status
@@ -75,17 +83,21 @@ esac;
 if [ "$user" == "custom" -o "$host" == "custom" ]; then
   if [ ! -z /tmp/anykernel/dtbo ]; then
     ui_print " "; ui_print "You are on a custom ROM, patching dtbo to remove verity...";
-    # Temporarily block out all custom recovery binaries/libs
-    mv /sbin /sbin_tmp
-    # Unset library paths
-    OLD_LD_LIB=$LD_LIBRARY_PATH
-    OLD_LD_PRE=$LD_PRELOAD
-    unset LD_LIBRARY_PATH
-    unset LD_PRELOAD
+    if $in_recovery; then
+      # Temporarily block out all custom recovery binaries/libs
+      mv /sbin /sbin_tmp;
+      # Unset library paths
+      OLD_LD_LIB=$LD_LIBRARY_PATH;
+      OLD_LD_PRE=$LD_PRELOAD;
+      unset LD_LIBRARY_PATH;
+      unset LD_PRELOAD;
+    fi;
     $bin/magiskboot --dtb-patch /tmp/anykernel/dtbo;
-    mv /sbin_tmp /sbin 2>/dev/null
-    [ -z $OLD_LD_LIB ] || export LD_LIBRARY_PATH=$OLD_LD_LIB
-    [ -z $OLD_LD_PRE ] || export LD_PRELOAD=$OLD_LD_PRE
+    if $in_recovery; then
+      mv /sbin_tmp /sbin 2>/dev/null;
+      [ -z $OLD_LD_LIB ] || export LD_LIBRARY_PATH=$OLD_LD_LIB;
+      [ -z $OLD_LD_PRE ] || export LD_PRELOAD=$OLD_LD_PRE;
+    fi;
   fi;
 else
   ui_print " "; ui_print "You are on stock, not patching dtbo to remove verity!";
@@ -93,10 +105,12 @@ fi;
 
 
 # Unmount system
-umount /system;
-umount /system_root;
-rmdir /system_root;
-mount -o ro -t auto /system;
+if $in_recovery; then
+  umount /system;
+  umount /system_root;
+  rmdir /system_root;
+  mount -o ro -t auto /system;
+fi;
 
 
 # Install the boot image
